WIP - reference all modules correctly
