+++ smart_news_query_embeddings/models/bert_keras_model.py
