+++ smart_news_query_embeddings/utils/model_utils.py
