+++ smart_news_query_embeddings/utils/utils.py
